requirements.txt
# Add Python packages here, one per line. Pin versions where possible.

# Snowflake provider for Airflow
apache-airflow-providers-snowflake

# Common SQL providers (used by Snowflake operators)
apache-airflow-providers-common-sql

# Additional data processing libraries for reporting
pandas